Refactored scripts to:  - break apart test data generation from actual test run  - added build + rebuild scenarios  - cleaned up page templates to remove usage of external components like 'Layout'
