+++ e2e/benchmark/.gitignore
@@ -1,0 +1,2 @@
+test-data*/
+*results.md
