+++ e2e/benchmark/pages/SG.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const ExamplePage = () => {
+  
+  return (
+    <>
+      <h1 className="my-5">Welcome to the Next Generation Entando Web Renderization Engine 👋</h1>
+    </>
+  );
+};
+
+export default ExamplePage;
