+++ e2e/benchmark/pages/SSG-single.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { Layout } from 'components/Layout';
 
 import productList from './data/products.json';
 
@@ -10,14 +9,12 @@ interface Props {
 const ExamplePage = ({ pageId }: Props) => {
   
   return (
-    <Layout title="Test Page">
-      <>
-        <h1 className="my-5">Welcome to the Next Generation Entando Web Renderization Engine 👋</h1>
-        <div className="mb-5 lead text-muted">
-          {pageId}
-        </div>
-      </>
-    </Layout>
+    <>
+      <h1 className="my-5">Welcome to the Next Generation Entando Web Renderization Engine 👋</h1>
+      <div className="mb-5 lead text-muted">
+        {pageId}
+      </div>
+    </>
   );
 };
 
@@ -33,7 +30,7 @@ export async function getStaticProps({ params }) {
 
 export async function getStaticPaths() {
 
-  const paths = productList.map(({name}) => ({ params: {product: name}}));
+  const paths = productList.map(({ name }) => ({ params: { product: name }}));
 
   return {
     paths,
