+++ e2e/benchmark/pages/data/products.json
