+++ starters/barebone-nextjs-engine/.eslintrc
@@ -25,7 +25,7 @@
     ],
     "settings": {
         "react": {
-            "version": "latest"
+            "version": "detect"
         }
     },
     "ignorePatterns": [ "public/*", "jest.config.*" ],
