--- starters/quickstart-nextjs-starter/benchmark/build.logs
@@ -1,3 +1,0 @@
-warn  - You have enabled experimental feature(s).
-warn  - Experimental features are not covered by semver, and may cause unexpected or broken application behavior. Use them at your own risk.
-
