--- starters/quickstart-nextjs-starter/benchmark/pages/SG.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Layout } from 'components/Layout';
-
-const ExamplePage = () => {
-  
-  return (
-    <Layout title="Test Page">
-      <>
-        <h1 className="my-5">Welcome to the Next Generation Entando Web Renderization Engine 👋</h1>
-      </>
-    </Layout>
-  );
-};
-
-export default ExamplePage;
