--- starters/quickstart-nextjs-starter/benchmark/results.md
@@ -1,36 +1,0 @@
-# BENCHMARK TEST
-## SSR Test Results
-Generating time: 2  
-Build time: 55  
-Folder size
-166M	../.next
- 
----
-## SSG Test Results
-Generating time: 2  
-Build time: 54  
-Folder size: 
-177M	../.next
- 
----
-## SG Test Results
-Generating time: 3  
-Build time: 45  
-Folder size: 
-154M	../.next
- 
----
-## SSG - Single page Test Results
-Generating time: 0  
-Build time: 15  
-Folder size: 
- 24M	../.next
- 
----
- 
- 
-## FINAL RESULTS
--------------------------
-- Build time: 169 seconds  
-- Generation time: 7 seconds  
-- Total time: 180 seconds
